+++ public/export.php
@@ -9,7 +9,6 @@ require '../vendor/autoload.php';
 use diskover\Constants;
 error_reporting(E_ALL ^ E_NOTICE);
 require "../src/diskover/Diskover.php";
-require "vars_inc.php";
 
 
 // Get search results from Elasticsearch